rewrite to vitest
