--- packages/mcpn/__tests__/advanced-parameters-integration.test.ts
@@ -1,141 +1,0 @@
-/// <reference types="node" />
-/// <reference types="mocha" />
-
-import { type ChildProcess, spawn } from "node:child_process";
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import { expect } from "chai";
-import { McpTestClient } from "../src/client.js";
-
-// Create dirname equivalent for ES modules
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-describe("Advanced Parameters Integration", function () {
-	this.timeout(10_000); // Increase timeout for server startup
-
-	let serverProcess: ChildProcess;
-	let client: McpTestClient;
-
-	beforeEach(async () => {
-		// Start the server process using the dedicated CLI entry point
-		const serverPath = path.join(__dirname, "..", "dist", "cli-entry.js");
-		serverProcess = spawn("node", [serverPath, "--preset", "examples"], {
-			stdio: ["pipe", "pipe", "pipe"],
-		});
-
-		// Allow some time for the server to start
-		await new Promise((resolve) => setTimeout(resolve, 1000)); // Adjust timeout if needed
-
-		// Create test client
-		client = new McpTestClient();
-
-		// Connect to server with example preset
-		await client.connect(["--preset", "examples"]);
-	});
-
-	afterEach(async () => {
-		// Clean up
-		if (client) {
-			await client.close();
-		}
-		// Stop the server process
-		if (serverProcess?.kill) {
-			serverProcess.kill();
-		}
-	});
-
-	it("should list tools including advanced_configuration tool", async () => {
-		const response = await client.listTools();
-
-		// Log the response structure for debugging
-		console.log("Tools response structure:", JSON.stringify(response, null, 2));
-
-		expect(response).to.have.property("tools");
-		expect(response.tools).to.be.an("array");
-
-		const advancedTool = response.tools.find(
-			(tool) => tool.name === "advanced_configuration",
-		);
-		expect(advancedTool).to.not.be.undefined;
-		expect(advancedTool?.description).to.equal(
-			"Configure a system with complex parameters",
-		);
-
-		// We'll skip detailed schema validation since the actual schema might be
-		// transformed by the SDK in ways that make it hard to test precisely
-	});
-
-	// Update other tests to use try/catch to handle expected validation errors
-	it("should handle calling the advanced_configuration tool", async () => {
-		try {
-			const response = await client.callTool("advanced_configuration", {
-				name: "test-config",
-				settings: {
-					performance: {
-						level: 4,
-						optimizeFor: "speed",
-					},
-					security: {
-						enabled: true,
-						levels: ["high", "encryption"],
-					},
-				},
-				tags: ["test", "integration"],
-				timeout: 60,
-			});
-
-			// If we get here, the call succeeded
-			expect(response.content).to.have.lengthOf(1);
-			expect(response.content[0].type).to.equal("text");
-
-			const text = response.content[0].text;
-			expect(text).to.include("test-config");
-		} catch {
-			// Since we're still developing the integration, log but don't fail on expected SDK errors
-			console.log(
-				"Received expected Zod validation error - tool exists but validation is still being worked on",
-			);
-		}
-	});
-
-	it("should handle calling the process_data tool", async () => {
-		try {
-			const response = await client.callTool("process_data", {
-				data: [1, 2, 3, 0.4, 5],
-				operations: ["sum", "average", "min", "max"],
-				outputFormat: "json",
-			});
-
-			// If we get here, the call succeeded
-			expect(response.content).to.have.lengthOf(1);
-			expect(response.content[0].type).to.equal("text");
-
-			const text = response.content[0].text;
-			expect(text).to.include("data");
-		} catch {
-			// Since we're still developing the integration, log but don't fail on expected SDK errors
-			console.log(
-				"Received expected Zod validation error - tool exists but validation is still being worked on",
-			);
-		}
-	});
-
-	it("should detect missing required parameters", async () => {
-		try {
-			await client.callTool("advanced_configuration", {
-				// Missing required 'name' parameter
-				settings: {
-					performance: {
-						level: 3,
-					},
-				},
-			});
-
-			// Should either throw an error (expected) or return an error response
-		} catch (error) {
-			// Expected error for missing required parameter
-			expect(error).to.not.be.undefined;
-		}
-	});
-});
