--- packages/mcpn/__tests__/test-workflows/.workflows/custom-name.yaml
@@ -1,7 +1,0 @@
-debugger_mode:
-  name: "custom_debugger"
-  description: "A debugger tool with a custom name"
-  prompt: |
-    # Custom Named Debugger
-
-    This debugger has a custom name to avoid conflicts with other MCP servers.
