--- packages/mcpn/__tests__/test-workflows/.workflows/disable-tool.yaml
@@ -1,2 +1,0 @@
-planner_mode:
-  disabled: true
