--- packages/mcpn/__tests__/test-workflows/.workflows/empty-tool.yaml
@@ -1,2 +1,0 @@
-empty_tool:
-  description: "Tool with minimal configuration"
