--- packages/mcpn/__tests__/test-workflows/.workflows/many-tools.yaml
@@ -1,26 +1,0 @@
-many_tools:
-  description: "Tool with many sub-tools"
-  prompt: |
-    # Many Tools Example
-  tools:
-    - name: "tool1"
-      description: "Tool 1"
-    - name: "tool2"
-      description: "Tool 2"
-    - name: "tool3"
-      description: "Tool 3"
-    - name: "tool4"
-      description: "Tool 4"
-    - name: "tool5"
-      description: "Tool 5"
-    - name: "tool6"
-      description: "Tool 6"
-    - name: "tool7"
-      description: "Tool 7"
-    - name: "tool8"
-      description: "Tool 8"
-    - name: "tool9"
-      description: "Tool 9"
-    - name: "tool10"
-      description: "Tool 10"
-    # This would continue with many more tools in a real test file
