--- packages/mcpn/__tests__/test-workflows/.workflows/optional-description.yaml
@@ -1,11 +1,0 @@
-optional_description_mode:
-  description: "Mode to test optional tool descriptions"
-  prompt: |
-    # Optional Description Test Mode
-    This mode is used to test tools with and without descriptions.
-  tools:
-    - name: "toolWithDescription"
-      description: "This tool has a description"
-    - name: "toolWithoutDescription"
-    - name: "anotherToolWithDescription"
-      description: "Another tool with a description"
