--- packages/mcpn/__tests__/test-workflows/.workflows/override-description.yaml
@@ -1,2 +1,0 @@
-debugger_mode:
-  description: "Custom debugging tool description"
