--- packages/mcpn/__tests__/test-workflows/.workflows/override-prompt.yaml
@@ -1,5 +1,0 @@
-debugger_mode:
-  prompt: |
-    # Custom Debugger Mode
-
-    This is a completely custom prompt for the debugger mode.
