--- packages/mcpn/__tests__/test-workflows/.workflows/override.yaml
@@ -1,15 +1,0 @@
-debugger_mode:
-  description: "A debugger tool with a custom name"
-  context: "This is a test override for the debugger mode."
-
-custom_tool:
-  description: "Custom test tool"
-  prompt: |
-    # Test Tool
-
-    This is a completely custom tool for testing purposes.
-
-    ## Test Steps
-    1. Step one
-    2. Step two
-    3. Step three
