--- packages/mcpn/__tests__/test-workflows/.workflows/sequential-tools.yaml
@@ -1,14 +1,0 @@
-sequential_tool:
-  description: "Tool with sequential sub-tools"
-  prompt: |
-    # Sequential Tool
-
-    This tool demonstrates sequential tool usage.
-
-  tools:
-    - name: "step1"
-      description: "First step in the sequence"
-    - name: "step2"
-      description: "Second step in the sequence"
-
-  toolMode: "sequential"
