--- packages/mcpn/__tests__/test-workflows/.workflows/special-chars.yaml
@@ -1,6 +1,0 @@
-special_chars_tool:
-  description: "Tool with special characters: !@#$%^&*()"
-  prompt: |
-    # Special Characters
-
-    This tool has `code`, *formatting*, and [links](https://example.com).
