--- packages/mcpn/__tests__/test-workflows/.workflows/tools-example.yaml
@@ -1,14 +1,0 @@
-example_tool:
-  description: "Example tool with sub-tools"
-  prompt: |
-    # Example Tool with Sub-Tools
-
-    This is an example tool that uses the 'tools' property.
-
-  tools:
-    - name: "subtool1"
-      description: "First sub-tool for testing"
-    - name: "subtool2"
-      description: "Second sub-tool for testing"
-
-  toolMode: "sequential"
