--- packages/mcpn/__tests__/test-workflows/.workflows/unicode.yaml
@@ -1,6 +1,0 @@
-unicode_tool:
-  description: "Tool with Unicode characters: 你好, 안녕하세요, Привет"
-  prompt: |
-    # Unicode Test
-
-    Unicode text: 你好, 안녕하세요, Привет, اَلْعَرَبِيَّةُ
