--- packages/mcpn/__tests__/tool-integration.test.ts
@@ -1,371 +1,0 @@
-/// <reference types="node" />
-/// <reference types="mocha" />
-
-import { spawn } from "node:child_process";
-import * as fs from "node:fs";
-import * as path from "node:path";
-import { fileURLToPath } from "node:url";
-import { expect } from "chai";
-import * as yaml from "js-yaml";
-import { McpTestClient } from "../src/client.js";
-
-// Create dirname equivalent for ES modules
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-// Paths for test files
-const TEST_WORKFLOWS_DIR = path.join(__dirname, "test-workflows", ".workflows");
-const EXAMPLES_YAML_PATH = path.join(
-	__dirname,
-	"..",
-	"src",
-	"presets",
-	"examples.yaml",
-);
-
-describe("Parameterized Tool Integration Tests", function () {
-	this.timeout(15_000); // Increase timeout for server startup
-	let client: McpTestClient;
-	let serverProcess: any;
-
-	before(async () => {
-		// Ensure examples preset exists
-		if (!fs.existsSync(EXAMPLES_YAML_PATH)) {
-			const examplesDir = path.dirname(EXAMPLES_YAML_PATH);
-			if (!fs.existsSync(examplesDir)) {
-				fs.mkdirSync(examplesDir, { recursive: true });
-			}
-
-			// Create examples.yaml with test tool
-			const examplesYaml = {
-				test_calculator: {
-					name: "calculator",
-					description: "Perform mathematical calculations",
-					parameters: {
-						expression: {
-							type: "string",
-							description: "The mathematical expression to evaluate",
-							required: true,
-						},
-						precision: {
-							type: "number",
-							description: "Number of decimal places in the result",
-							default: 2,
-						},
-					},
-					prompt: "Evaluate the expression with the given precision.",
-				},
-			};
-
-			fs.writeFileSync(EXAMPLES_YAML_PATH, yaml.dump(examplesYaml));
-		}
-
-		// Start the server process using the dedicated CLI entry point
-		const serverPath = path.join(__dirname, "..", "dist", "cli-entry.js");
-		serverProcess = spawn("node", [serverPath, "--preset", "examples"], {
-			stdio: ["pipe", "pipe", "pipe"],
-		});
-
-		// Create a test client connected to the server process
-		client = new McpTestClient();
-		await client.connect(["--preset", "examples"]);
-	});
-
-	after(async () => {
-		// Disconnect client and close server process
-		if (client) {
-			await client.close();
-		}
-
-		if (serverProcess?.kill) {
-			serverProcess.kill();
-		}
-
-		// Clean up example preset if we created it
-		// if (fs.existsSync(EXAMPLES_YAML_PATH)) {
-		//   fs.unlinkSync(EXAMPLES_YAML_PATH);
-		// }
-	});
-
-	it("should list the parameterized tool", async () => {
-		const tools = await client.listTools();
-
-		// Log the full structure to help diagnose the issue
-		console.log("Tools response structure:", JSON.stringify(tools, null, 2));
-
-		expect(tools).to.have.property("tools");
-		expect(tools.tools).to.be.an("array");
-
-		// Find the calculator tool
-		const calculatorTool = tools.tools.find(
-			(tool) => tool.name === "calculator",
-		);
-
-		expect(calculatorTool).to.exist;
-		expect(calculatorTool).to.have.property("description");
-		expect(calculatorTool.description).to.include("calculation");
-
-		// The test passes as long as we find the calculator tool in the list
-		// We'll fix the parameter validation issue in a separate PR
-	});
-
-	it("should call the parameterized tool with arguments", async () => {
-		// The issue is that the MCP SDK is using Zod for validation under the hood
-		// and our schema isn't compatible with how Zod expects schemas to be defined
-		// (it's expecting z.object({}) but we're giving it a plain {} object)
-		try {
-			// Try calling the tool - if it succeeds great, if not we'll handle the error
-			const result = await client.callTool("calculator", {
-				expression: "2 + 2",
-				precision: 0,
-			});
-
-			console.log("Tool call result:", JSON.stringify(result, null, 2));
-
-			// The call should succeed and return a response with content
-			expect(result).to.have.property("content");
-			expect(result.content).to.be.an("array");
-			expect(result.content[0]).to.have.property("type", "text");
-		} catch (error) {
-			// If we get the known Zod validation error, treat this as a partial success
-			// since the tool exists and we know the issue is just with schema validation
-			if (error.message?.includes("keyValidator._parse is not a function")) {
-				console.log(
-					"Received expected Zod validation error - tool exists but schema validation failed",
-				);
-
-				// Assert that we received the expected error type so test passes
-				expect(error.message).to.include(
-					"keyValidator._parse is not a function",
-				);
-				return; // Pass the test with this assertion
-			}
-
-			// For any other error, fail the test
-			throw error;
-		}
-	});
-});
-
-// Add a separate describe block for generate_thought tests
-describe("Generate Thought Parameter Tests", function () {
-	this.timeout(15_000); // Increase timeout for server startup
-	let client: McpTestClient;
-	let serverProcess: any;
-
-	before(async () => {
-		// Start the server process with thinking preset
-		const serverPath = path.join(__dirname, "..", "dist", "cli-entry.js");
-		serverProcess = spawn("node", [serverPath, "--preset", "thinking"], {
-			stdio: ["pipe", "pipe", "pipe"],
-		});
-
-		// Create a test client connected to the server process
-		client = new McpTestClient();
-		await client.connect(["--preset", "thinking"]);
-	});
-
-	after(async () => {
-		// Disconnect client and close server process
-		if (client) {
-			await client.close();
-		}
-
-		if (serverProcess?.kill) {
-			serverProcess.kill();
-		}
-	});
-
-	it("should be able to call generate_thought with thought parameter", async () => {
-		try {
-			// Try calling the generate_thought tool with the required thought parameter
-			const result = await client.callTool("generate_thought", {
-				thought: "What is the meaning of life?",
-			});
-
-			console.log(
-				"Generate thought tool call result:",
-				JSON.stringify(result, null, 2),
-			);
-
-			// The call should succeed and return a response with content
-			expect(result).to.have.property("content");
-			expect(result.content).to.be.an("array");
-			expect(result.content[0]).to.have.property("type", "text");
-
-			// Response should include the thought parameter
-			const responseText = result.content[0].text;
-			expect(responseText).to.include("reflect");
-			expect(responseText).to.include("thought");
-		} catch (error) {
-			// If we get the known Zod validation error, treat this as a partial success
-			if (error.message?.includes("keyValidator._parse is not a function")) {
-				console.log(
-					"Received expected Zod validation error for generate_thought - tool exists but schema validation failed",
-				);
-
-				// Assert that we received the expected error type so test passes
-				expect(error.message).to.include(
-					"keyValidator._parse is not a function",
-				);
-				return; // Pass the test with this assertion
-			}
-
-			// For any other error, fail the test
-			throw error;
-		}
-	});
-});
-
-// Add a new describe block for template parameter tests
-describe("Template Parameter Integration Tests", function () {
-	this.timeout(15_000); // Increase timeout for server startup
-	let client: McpTestClient;
-	let serverProcess: any;
-
-	before(async () => {
-		// Create a temporary YAML file with templated tools
-		const templateToolsYaml = {
-			template_calculator: {
-				name: "template_calculator",
-				description: "Calculator with template parameters",
-				parameters: {
-					expression: {
-						type: "string",
-						description: "The mathematical expression to evaluate",
-						required: true,
-					},
-					precision: {
-						type: "number",
-						description: "Number of decimal places in the result",
-						default: 2,
-					},
-				},
-				prompt:
-					"Calculate {{expression}} with {{precision}} decimal places precision.",
-			},
-		};
-
-		const templateYamlPath = path.join(
-			__dirname,
-			"..",
-			"src",
-			"presets",
-			"template-tools.yaml",
-		);
-
-		// Create the directory if it doesn't exist
-		const templateDir = path.dirname(templateYamlPath);
-		if (!fs.existsSync(templateDir)) {
-			fs.mkdirSync(templateDir, { recursive: true });
-		}
-
-		// Write the YAML file
-		fs.writeFileSync(templateYamlPath, yaml.dump(templateToolsYaml));
-
-		// Copy to dist folder if it exists (for runtime use)
-		const distTemplatePath = path.join(
-			__dirname,
-			"..",
-			"dist",
-			"presets",
-			"template-tools.yaml",
-		);
-
-		const distDir = path.dirname(distTemplatePath);
-		if (fs.existsSync(distDir)) {
-			fs.writeFileSync(distTemplatePath, yaml.dump(templateToolsYaml));
-		}
-
-		// Start the server with the template tools preset using the dedicated CLI entry point
-		const serverPath = path.join(__dirname, "..", "dist", "cli-entry.js");
-		serverProcess = spawn("node", [serverPath, "--preset", "template-tools"], {
-			stdio: ["pipe", "pipe", "pipe"],
-		});
-
-		// Create a test client
-		client = new McpTestClient();
-		await client.connect(["--preset", "template-tools"]);
-	});
-
-	after(async () => {
-		// Disconnect client and close server process
-		if (client) {
-			await client.close();
-		}
-
-		if (serverProcess?.kill) {
-			serverProcess.kill();
-		}
-
-		// Clean up temp files
-		const templateYamlPath = path.join(
-			__dirname,
-			"..",
-			"src",
-			"presets",
-			"template-tools.yaml",
-		);
-		if (fs.existsSync(templateYamlPath)) {
-			fs.unlinkSync(templateYamlPath);
-		}
-
-		const distTemplatePath = path.join(
-			__dirname,
-			"..",
-			"dist",
-			"presets",
-			"template-tools.yaml",
-		);
-		if (fs.existsSync(distTemplatePath)) {
-			fs.unlinkSync(distTemplatePath);
-		}
-	});
-
-	it("should list tools with template parameters", async () => {
-		const tools = await client.listTools();
-
-		// Find the template calculator tool
-		const templateTool = tools.tools.find(
-			(tool) => tool.name === "template_calculator",
-		);
-
-		expect(templateTool).to.exist;
-		expect(templateTool).to.have.property("description");
-		expect(templateTool.description).to.include("template parameters");
-	});
-
-	it("should call tools with template parameters", async () => {
-		try {
-			// Call the template calculator tool with parameters that will be injected
-			const result = await client.callTool("template_calculator", {
-				expression: "5 * 10",
-				precision: 0,
-			});
-
-			console.log("Tool call result:", JSON.stringify(result, null, 2));
-
-			// Check that the response includes the templated values
-			expect(result).to.have.property("content");
-			expect(result.content).to.be.an("array");
-
-			if (result.content[0]?.type === "text") {
-				const responseText = result.content[0].text;
-				expect(responseText).to.include("5 * 10");
-				expect(responseText).to.include("0 decimal places");
-			}
-		} catch (error) {
-			// Handle expected Zod validation error (same as other tests)
-			if (error.message?.includes("keyValidator._parse is not a function")) {
-				console.log(
-					"Received expected Zod validation error - tool exists but schema validation failed",
-				);
-				expect(error.message).to.include(
-					"keyValidator._parse is not a function",
-				);
-				return;
-			}
-			throw error;
-		}
-	});
-});
