+++ packages/mcpn/playground/package.json
@@ -1,0 +1,19 @@
+{
+  "private": true,
+  "name": "mcpn-playground",
+  "type": "module",
+  "scripts": {
+    "dev": "nuxi dev",
+    "build": "nuxi build",
+    "generate": "nuxi generate",
+    "test:types": "vue-tsc --noEmit"
+  },
+  "dependencies": {
+    "@nuxt/fonts": "latest",
+    "@nuxtjs/tailwindcss": "latest",
+    "nuxt": "latest",
+    "tailwindcss": "^3.4.17",
+    "vue": "3.5.13",
+    "vue-router": "latest"
+  }
+}
