+++ packages/mcpn/test/test-workflows/.workflows/custom-tool.yaml
@@ -3,4 +3,4 @@ custom_tool:
   prompt: |
     # Custom Tool
 
-    This is a custom tool that doesn't exist in presets.
+	This is a custom tool that doesn't exist in presets.
