+++ packages/mcpn/test/test-workflows/.workflows/disable-tool.yaml
@@ -1,0 +1,2 @@
+planner_mode:
+	disabled: true
