+++ packages/mcpn/test/test-workflows/.workflows/override-description.yaml
@@ -1,0 +1,2 @@
+debugger_mode:
+	description: "Custom debugging tool description"
