+++ packages/mcpn/test/test-workflows/.workflows/override-prompt.yaml
@@ -1,0 +1,5 @@
+debugger_mode:
+  prompt: |
+    # Custom Debugger Mode
+
+	This is a completely custom prompt for the debugger mode.
