+++ packages/mcpn/vitest.config.mjs
@@ -1,0 +1,10 @@
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+	test: {
+		coverage: {
+			include: ["src/**/*.ts"],
+			exclude: ["src/@types/**"],
+		},
+	},
+});
